+++ CHANGELOG.md
@@ -198,6 +198,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * A nearly identical string pattern system as revsets is now supported in the
   template language, and is exposed as `string.match(pattern)`.
 
+* Merge tools can use the `$path` argument to learn where the file they
+  are merging will end up in the repository.
+
 ### Fixed bugs
 
 * `jj git clone` now correctly fetches all tags, unless `--fetch-tags` is